@@ -13,10 +13,10 @@
   "license": "ISC",
   "devDependencies": {
     "@babel/core": "^7.11.6",
-    "@babel/preset-react": "^7.13.13",
     "@babel/plugin-proposal-class-properties": "^7.10.4",
     "@babel/plugin-transform-runtime": "^7.10.1",
     "@babel/preset-env": "^7.11.5",
+    "@babel/preset-react": "^7.13.13",
     "babel-loader": "^8.1.0",
     "clean-webpack-plugin": "^3.0.0",
     "css-loader": "^4.2.2",
@@ -31,11 +31,10 @@
   "dependencies": {
     "bootstrap": "^4.5.2",
     "cross-env": "^7.0.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-bootstrap": "^1.3.0",
-    "react-router": "^5.2.0",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.2.0"
+    "react": "^18.1.0",
+    "react-bootstrap": "^2.4.0",
+    "react-dom": "^18.1.0",
+    "react-router-bootstrap": "^0.26.1",
+    "react-router-dom": "^6.3.0"
   }
 }
